requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# AI/ML dependencies for Business Insights AI
langgraph>=0.6.6
langchain-openai>=0.3.32
langsmith>=0.4.21
pandas>=2.3.2
numpy>=2.3.2
matplotlib>=3.10.6
seaborn>=0.13.2
openpyxl>=3.1.5
python-multipart>=0.0.20
requests>=2.32.5
